make tn-engine it tests
